Make Basic Authentication Optional for Custom Metrics

+++ api/broker/broker.go
@@ -47,6 +47,7 @@ var (
 	ErrDeletePolicyForUnbinding    = errors.New("failed to delete policy for unbinding")
 	ErrDeleteServiceBinding        = errors.New("error deleting service binding")
 	ErrCredentialNotDeleted        = errors.New("failed to delete custom metrics credential for unbinding")
+	ErrInvalidCredentialType       = errors.New("invalid credential type provided: allowed values are [binding-secret, x509]")
 )
 
 type Errors []error
@@ -510,6 +511,11 @@ func (b *Broker) Bind(ctx context.Context, instanceID string, bindingID string, 
 		logger.Error("get-default-policy", err)
 		return result, err
 	}
+	defaultCustomMetricsCredentialType := b.conf.DefaultCustomMetricsCredentialType
+	customMetricsBindingAuthScheme, err := getOrDefaultCredentialType(policyJson, defaultCustomMetricsCredentialType, logger)
+	if err != nil {
+		return result, err
+	}
 	policyGuid, err := uuid.NewV4()
 	if err != nil {
 		b.logger.Error("get-default-policy-create-guid", err, lager.Data{"instanceID": instanceID})
@@ -547,18 +553,33 @@ func (b *Broker) Bind(ctx context.Context, instanceID string, bindingID string, 
 		}
 		return result, apiresponses.NewFailureResponse(ErrCreatingServiceBinding, http.StatusInternalServerError, actionCreateServiceBinding)
 	}
+	customMetricsCredentials := &models.CustomMetricsCredentials{
+		MtlsUrl: b.conf.MetricsForwarder.MetricsForwarderMtlsUrl,
+	}
+	if !isValidCredentialType(customMetricsBindingAuthScheme.CredentialType) {
+		actionValidateCredentialType := "validate-credential-type" // #nosec G101
+		logger.Error("invalid credential-type provided", err, lager.Data{"credential-type": customMetricsBindingAuthScheme.CredentialType})
+		return result, apiresponses.NewFailureResponseBuilder(
+			ErrInvalidCredentialType, http.StatusBadRequest, actionValidateCredentialType).
+			WithErrorKey(actionValidateCredentialType).
+			Build()
+	}
 
-	// create credentials
-	cred, err := b.credentials.Create(ctx, appGUID, nil)
-	if err != nil {
-		//revert binding creating
-		logger.Error("create-credentials", err)
-
-		err = b.bindingdb.DeleteServiceBindingByAppId(ctx, appGUID)
+	if customMetricsBindingAuthScheme.CredentialType == models.BindingSecret {
+		// create credentials
+		cred, err := b.credentials.Create(ctx, appGUID, nil)
 		if err != nil {
-			logger.Error("revert-binding-creation-due-to-credentials-creation-failure", err)
+			//revert binding creating
+			logger.Error("create-credentials", err)
+
+			err = b.bindingdb.DeleteServiceBindingByAppId(ctx, appGUID)
+			if err != nil {
+				logger.Error("revert-binding-creation-due-to-credentials-creation-failure", err)
+			}
+			return result, apiresponses.NewFailureResponse(ErrCreatingServiceBinding, http.StatusInternalServerError, "revert-binding-creation-due-to-credentials-creation-failure")
 		}
-		return result, apiresponses.NewFailureResponse(ErrCreatingServiceBinding, http.StatusInternalServerError, "revert-binding-creation-due-to-credentials-creation-failure")
+		customMetricsCredentials.URL = &b.conf.MetricsForwarder.MetricsForwarderUrl
+		customMetricsCredentials.Credential = cred
 	}
 
 	// attach policy to appGUID
@@ -567,14 +588,22 @@ func (b *Broker) Bind(ctx context.Context, instanceID string, bindingID string, 
 	}
 
 	result.Credentials = models.Credentials{
-		CustomMetrics: models.CustomMetricsCredentials{
-			Credential: cred,
-			URL:        b.conf.MetricsForwarder.MetricsForwarderUrl,
-			MtlsUrl:    b.conf.MetricsForwarder.MetricsForwarderMtlsUrl,
-		},
+		CustomMetrics: *customMetricsCredentials,
 	}
-
 	return result, nil
+}
+
+func getOrDefaultCredentialType(policyJson json.RawMessage, credentialTypeConfig string, logger lager.Logger) (*models.CustomMetricsBindingAuthScheme, error) {
+	credentialType := &models.CustomMetricsBindingAuthScheme{CredentialType: credentialTypeConfig}
+	if len(policyJson) != 0 {
+		err := json.Unmarshal(policyJson, &credentialType)
+		if err != nil {
+			logger.Error("error: unmarshal-credential-type", err)
+			return nil, apiresponses.NewFailureResponse(ErrCreatingServiceBinding, http.StatusInternalServerError, "error-unmarshal-credential-type")
+		}
+	}
+	logger.Debug("getOrDefaultCredentialType", lager.Data{"credential-type": credentialType})
+	return credentialType, nil
 }
 
 func (b *Broker) attachPolicyToApp(ctx context.Context, appGUID string, policy *models.ScalingPolicy, policyGuidStr string, logger lager.Logger) error {
@@ -824,4 +853,7 @@ func (b *Broker) deleteBinding(ctx context.Context, bindingId string, serviceIns
 		return ErrCredentialNotDeleted
 	}
 	return nil
+}
+func isValidCredentialType(credentialType string) bool {
+	return credentialType == models.BindingSecret || credentialType == models.X509Certificate
 }
